extracted UI inputs out of battle-scene
