+++ src/ui/menu-ui-handler.ts
@@ -9,7 +9,7 @@ import { OptionSelectConfig, OptionSelectItem } from "./abstact-option-select-ui
 import { Tutorial, handleTutorial } from "../tutorial";
 import { updateUserInfo } from "../account";
 import i18next from '../plugins/i18n';
-import {Button} from "#app/inputsController";
+import {Button} from "#app/inputs-controller";
 
 export enum MenuOptions {
   GAME_SETTINGS,
