+++ src/ui/summary-ui-handler.ts
@@ -17,7 +17,7 @@ import { Nature, getNatureStatMultiplier } from "../data/nature";
 import { loggedInUser } from "../account";
 import { PlayerGender } from "../system/game-data";
 import { Variant, getVariantTint } from "#app/data/variant";
-import {Button} from "#app/inputsController";
+import {Button} from "#app/inputs-controller";
 
 enum Page {
   PROFILE,
